+++ R/log.R
@@ -5,11 +5,13 @@ read_info <- function(
   md,
   start,
   log,
+  pkgs_used,
   session,
   environment,
   options,
-  python = NULL,
-  approved_packages = NULL,
+  python_pip_list = NULL,
+  python_new_status = NULL,
+  python_old_status = NULL,
   track_files = FALSE
 ) {
   info <- list(
@@ -20,14 +22,14 @@ read_info <- function(
       split_log(),
     session = read_session_info(
       file = session,
-      approved_packages = approved_packages
+      pkgs_used = pkgs_used
     )
   )
 
   info$session$environment <- read_environment(environment)
   info$session$options <- read_options(options)
 
-  if (!is.null(python) && file.exists(python)) {
+  if (!is.null(python_pip_list) && file.exists(python_pip_list)) {
     info$session$platform <- info$session$platform |>
       dplyr::bind_rows(
         data.frame(
@@ -35,7 +37,13 @@ read_info <- function(
           value = python_version()
         )
       )
-    info$session$python <- read_python(python)
+
+    info$session$python <- read_python(
+      old_status = python_old_status,
+      new_status = python_new_status,
+      pip_list = python_pip_list
+    )
+
     info$session <-
       info$session[c("platform", "R", "python", "environment", "options")]
   }
@@ -45,8 +53,9 @@ read_info <- function(
 
 #' Read and format session info output from `sessioninfo::session_info()`
 #' @noRd
-read_session_info <- function(file, approved_packages = NULL) {
+read_session_info <- function(file, pkgs_used) {
   info <- readRDS(file)
+  pkgs_used <- readRDS(pkgs_used)
 
   platform <- info[["platform"]] |>
     unlist() |>
@@ -58,10 +67,6 @@ read_session_info <- function(file, approved_packages = NULL) {
       package = .data$package,
       version = .data$loadedversion,
       attached = .data$attached,
-      approved = check_approved(
-        used = paste(.data$package, .data$version, sep = "@"),
-        approved = approved_packages
-      ),
       path = .data$loadedpath,
       date = vapply(
         X = .data$package,
@@ -79,12 +84,18 @@ read_session_info <- function(file, approved_packages = NULL) {
         },
         FUN.VALUE = character(1),
         USE.NAMES = FALSE
-      )
+      ),
+      directly_used = .data$package %in% pkgs_used[["Package"]],
+      approved = check_approved(
+        used = paste(.data$package, .data$version, sep = "@"),
+        approved = zephyr::get_option("approved_packages")
+      ),
     ) |>
     dplyr::select(
       "package",
       "version",
       "attached",
+      "directly_used",
       "approved",
       "path",
       "date",
@@ -108,25 +119,13 @@ read_environment <- function(file) {
     dplyr::filter(
       stringr::str_detect(
         string = .data$variable,
-        pattern = paste0(r_secrets(), collapse = "|"),
+        pattern = paste0(
+          zephyr::get_option("environment_secrets", "whirl"),
+          collapse = "|"
+        ),
         negate = TRUE
       )
     )
-}
-
-#' List of patterns naming a secret environment variable
-#' any match will not be included in the log
-#' @noRd
-r_secrets <- function() {
-  c(
-    "BASH_FUNC",
-    "_SSL_CERT",
-    "_KEY",
-    "_KEY",
-    "_KEY",
-    "_PAT",
-    "_TOKEN"
-  )
 }
 
 #' Read and format options output from `options()`
@@ -148,26 +147,62 @@ python_version <- function() {
 }
 
 #' Read and format python packages information from a JSON file
-#' JSON file created in `inst/documents/python_modules.py`
+#' JSON files created in `inst/documents/python_modules.py`.
+#' Pip list is created in `ìnst/documents/dummy.qmd`.
 #' @noRd
-read_python <- function(json) {
-  json <- jsonlite::fromJSON(json)
+read_python <- function(old_status, new_status, pip_list) {
+  old <- old_status |>
+    jsonlite::read_json() |>
+    lapply(FUN = unlist, use.names = FALSE)
 
-  if (!length(json)) {
+  new <- new_status |>
+    jsonlite::read_json() |>
+    lapply(FUN = unlist, use.names = FALSE)
+
+  pip <- pip_list |>
+    readRDS() |>
+    parse_pip_list()
+
+  if (!nrow(pip)) {
     return(
       tibble::tibble(
-        Package = character(0),
-        Version = character(0),
-        Path = character(0)
+        package = character(0),
+        version = character(0),
+        directly_used = logical(0),
+        approved = logical(0),
+        path = character(0)
       )
     )
   }
 
-  json |>
-    tibble::enframe(name = "Package") |>
-    tidyr::unnest_wider(col = "value") |>
-    dplyr::rename(
-      Version = "version",
-      Path = "installation_path"
-    )
+  pip |>
+    dplyr::filter(
+      .data$package %in%
+        c(
+          setdiff(new$namespaced, old$namespaced),
+          setdiff(new$loaded, old$loaded)
+        )
+    ) |>
+    dplyr::mutate(
+      directly_used = .data$package %in% new$namespaced,
+      approved = check_approved(
+        used = paste(.data$package, .data$version, sep = "@"),
+        approved = zephyr::get_option("approved_python_packages")
+      )
+    ) |>
+    dplyr::select("package", "version", "directly_used", "approved", "path")
+}
+
+#' @noRd
+parse_pip_list <- function(x) {
+  x <- utils::read.table(text = x)
+
+  nm <- tolower(x[1, ])
+  nm[which(nm == "location")] <- "path"
+
+  names(x) <- nm
+
+  x |>
+    utils::tail(-2) |>
+    tibble::as_tibble()
 }