+++ R/whirl-options.R
@@ -55,8 +55,7 @@ zephyr::create_option(
     "^/urandom",
     "^/.cache"
   ),
-  description =
-    "List of file naming patterns not be tracked when track_files = TRUE"
+  description = "List of file naming patterns not be tracked when track_files = TRUE"
 )
 
 zephyr::create_option(
@@ -69,7 +68,13 @@ zephyr::create_option(
 zephyr::create_option(
   name = "approved_packages",
   default = NULL,
-  description = "List of approved packages and their version in the format: \\{name\\}@\\{version\\}"
+  description = "List of approved R packages and their version in the format: \\{name\\}@\\{version\\}"
+)
+
+zephyr::create_option(
+  name = "approved_python_packages",
+  default = NULL,
+  description = "List of approved Python packages and their version in the format: \\{name\\}@\\{version\\}"
 )
 
 zephyr::create_option(
@@ -103,4 +108,17 @@ zephyr::create_option(
   default = 9000,
   description = "Timeout for waiting for the R process from callr::r_session to 
   start, in milliseconds."
+)
+
+zephyr::create_option(
+  name = "environment_secrets",
+  default = c(
+    "BASH_FUNC",
+    "_SSL_CERT",
+    "_KEY",
+    "_PAT",
+    "_TOKEN"
+  ),
+  description = "Secret environment variable patterns. 
+  Any variables matching will not be included in the logs."
 )