+++ R/whirl_queue.R
@@ -17,7 +17,6 @@ whirl_queue <- R6::R6Class(
     initialize = \(
       # jscpd:ignore-start
       n_workers = zephyr::get_option("n_workers", "whirl"),
-      verbosity_level = zephyr::get_option("verbosity_level", "whirl"),
       check_renv = zephyr::get_option("check_renv", "whirl"),
       track_files = zephyr::get_option("track_files", "whirl"),
       out_formats = zephyr::get_option("out_formats", "whirl"),
@@ -26,7 +25,6 @@ whirl_queue <- R6::R6Class(
         "whirl"
       ),
       track_files_keep = zephyr::get_option("track_files_keep", "whirl"),
-      approved_packages = zephyr::get_option("approved_packages", "whirl"),
       log_dir = zephyr::get_option("log_dir", "whirl")
       # jscpd:ignore-end
     ) {
@@ -34,13 +32,11 @@ whirl_queue <- R6::R6Class(
         self,
         private,
         n_workers,
-        verbosity_level,
         check_renv,
         track_files,
         out_formats,
         track_files_discards,
         track_files_keep,
-        approved_packages,
         log_dir
       )
     },
@@ -87,7 +83,7 @@ whirl_queue <- R6::R6Class(
     #' the queue
     #' @return [invisible] self
     run = \(scripts, tag = NA_character_) {
-      wq_run(scripts, tag, self)
+      wq_run(scripts, tag, self, private)
     },
 
     #' @description Print method displaying the current status of the queue
@@ -129,14 +125,13 @@ whirl_queue <- R6::R6Class(
     .queue = NULL,
     .workers = NULL,
     .n_workers = NULL,
-    verbosity_level = NULL,
     check_renv = NULL,
     track_files = NULL,
     out_formats = NULL,
     track_files_discards = NULL,
     track_files_keep = NULL,
-    approved_packages = NULL,
-    log_dir = NULL
+    log_dir = NULL,
+    progress_bar = NULL
   )
 )
 
@@ -144,22 +139,18 @@ wq_initialise <- function(
   self,
   private,
   n_workers,
-  verbosity_level,
   check_renv,
   track_files,
   out_formats,
   track_files_discards,
   track_files_keep,
-  approved_packages,
   log_dir
 ) {
   private$check_renv <- check_renv
-  private$verbosity_level <- verbosity_level
   private$track_files <- track_files
   private$out_formats <- out_formats
   private$track_files_discards <- track_files_discards
   private$track_files_keep <- track_files_keep
-  private$approved_packages <- approved_packages
   private$log_dir <- log_dir
 
   private$.queue <- tibble::tibble(
@@ -195,7 +186,7 @@ wq_add_queue <- function(self, private, scripts, tag, status) {
     # Check if the directory exists
     unique_folders <- unique(folder)
     if (any(!file.exists(unique_folders))) {
-      missing <- unique_folders[!file.exists(unique_folders)]  # nolint: object_usage_linter
+      missing <- unique_folders[!file.exists(unique_folders)] # nolint: object_usage_linter
       cli::cli_abort(
         "Logs cannot be saved because {.val {missing}} does not exist"
       )
@@ -236,12 +227,10 @@ wq_poll <- function(
       n = length(wid),
       expr = whirl_r_session$new(
         check_renv = private$check_renv,
-        verbosity_level = private$verbosity_level,
         track_files = private$track_files,
         out_formats = private$out_formats,
         track_files_discards = private$track_files_discards,
         track_files_keep = private$track_files_keep,
-        approved_packages = private$approved_packages,
         log_dir = private$log_dir
       ),
       simplify = FALSE
@@ -258,7 +247,9 @@ wq_poll <- function(
   i_timeout <- round(timeout / length(i_active))
   for (i in i_active) {
     p <- private$.workers$session[[i]]$poll(timeout = i_timeout)
-    if (p == "ready") private$.workers$session[[i]]$read()
+    if (p == "ready") {
+      private$.workers$session[[i]]$check_status()
+    }
     if (private$.workers$session[[i]]$get_state() == "idle") {
       wq_next_step(self, private, i)
     }
@@ -272,6 +263,7 @@ wq_wait <- function(self, private, timeout) {
   timeout <- timeout / 1000 # Convert to secs
   go <- TRUE
   while (go) {
+    pb_update(id = private$progress_bar, queue = self$queue)
     self$poll(50)
     go <- any(self$queue$status %in% c("waiting", "running"))
     if (timeout >= 0 && difftime(Sys.time(), start, units = "secs") > timeout) {
@@ -301,7 +293,7 @@ wq_next_step <- function(self, private, wid) {
     # Step 3: Finish log and create outputs
     "3" = {
       private$.queue$result[[id_script]] <-
-        session$log_finish()$create_outputs(
+        session$log_finish(
           out_dir = private$.queue$log_dir[[id_script]],
           format = private$out_formats
         )
@@ -319,7 +311,11 @@ wq_next_step <- function(self, private, wid) {
   return(invisible(wid))
 }
 
-wq_run <- function(scripts, tag, self) {
+wq_run <- function(scripts, tag, self, private) {
+  private$progress_bar <- pb_start()
+  on.exit({
+    private$progress_bar <- pb_done(id = private$progress_bar)
+    gc() # finalizes used whirl_r_sessions - cleanup temp folders
+  })
   self$push(scripts = scripts, tag = tag)$wait()
-  on.exit(gc()) # finalizes used whirl_r_sessions - cleanup temp folders
 }