+++ R/whirl_r_session.R
@@ -14,7 +14,6 @@ whirl_r_session <- R6::R6Class(
     #' @return A [whirl_r_session] object
     initialize = \(
       # jscpd:ignore-start
-      verbosity_level = zephyr::get_option("verbosity_level", "whirl"),
       check_renv = zephyr::get_option("check_renv", "whirl"),
       track_files = zephyr::get_option("track_files", "whirl"),
       out_formats = zephyr::get_option("out_formats", "whirl"),
@@ -23,19 +22,16 @@ whirl_r_session <- R6::R6Class(
         "whirl"
       ),
       track_files_keep = zephyr::get_option("track_files_keep", "whirl"),
-      approved_packages = zephyr::get_option("approved_packages", "whirl"),
       log_dir = zephyr::get_option("log_dir", "whirl"),
       wait_timeout = zephyr::get_option("wait_timeout", "whirl")
       # jscpd:ignore-end
     ) {
       wrs_initialize(
-        verbosity_level,
         check_renv,
         track_files,
         out_formats,
         track_files_discards,
         track_files_keep,
-        approved_packages,
         log_dir,
         wait_timeout,
         self,
@@ -70,20 +66,6 @@ whirl_r_session <- R6::R6Class(
       wrs_check_status(self, private, super)
     },
 
-    #' @description Update the progress bar
-    #' @param ... parsed to [cli::cli_progress_update()]
-    #' @return [invisible] self
-    pb_update = \(...) {
-      wrs_pb_update(..., self = self, private = private, super = super)
-    },
-
-    #' @description Finalise the progress bar
-    #' @param status Status of the script. success, warning, or error
-    #' @return [invisible] self
-    pb_done = \(status) {
-      wrs_pb_done(status, self, private, super)
-    },
-
     #' @description Log the script
     #' @param script The script to log
     #' @return [invisible] self
@@ -97,21 +79,15 @@ whirl_r_session <- R6::R6Class(
       wrs_create_log(self, private, super)
     },
 
-    #' @description Finish the log
-    #' @return [invisible] self
-    log_finish = \() {
-      wrs_log_finish(self, private, super)
-    },
-
-    #' @description Create all log outputs
+    #' @description Finish the log and create outputs
     #' @param out_dir [character] Output directory for the log
     #' @param format [character] Output formats to create
-    #' @return [invisible],[list] of logging information
-    create_outputs = \(
+    #' @return [invisible] self
+    log_finish = \(
       out_dir,
       format = zephyr::get_option("out_formats", "whirl")
     ) {
-      wrs_create_outputs(out_dir, format, self, private, super)
+      wrs_log_finish(out_dir, format, self, private, super)
     }
   ),
   private = list(
@@ -119,16 +95,13 @@ whirl_r_session <- R6::R6Class(
     finalize = \() {
       wrs_finalize(self, private, super)
     },
-    verbosity_level = NULL,
     wd = NULL,
     track_files = NULL,
     out_formats = NULL,
     log_dir = NULL,
     track_files_discards = NULL,
     track_files_keep = NULL,
-    approved_packages = NULL,
     check_renv = NULL,
-    pb = NULL,
     current_script = NULL,
     track_files_log = "log_msg.json",
     result = NULL
@@ -137,36 +110,26 @@ whirl_r_session <- R6::R6Class(
 )
 
 wrs_initialize <- function(
-  verbosity_level,
   check_renv,
   track_files,
   out_formats,
   track_files_discards,
   track_files_keep,
-  approved_packages,
   log_dir,
   wait_timeout,
   self,
   private,
   super
 ) {
-  super$initialize(wait_timeout = 9000) # uses callr::r_session$initialize()
+  super$initialize(wait_timeout = wait_timeout) # uses callr::r_session$initialize()
 
   private$wd <- withr::local_tempdir(clean = FALSE)
-  private$verbosity_level <- verbosity_level
   private$check_renv <- check_renv
   private$track_files <- track_files
   private$out_formats <- out_formats
   private$track_files_discards <- track_files_discards
   private$track_files_keep <- track_files_keep
-  private$approved_packages <- approved_packages
   private$log_dir <- log_dir
-
-  # If the stream does not support dynamic tty, which is needed for progress
-  # bars to update in place, the verbosity is downgraded.
-  if (private$verbosity_level == "verbose" && !cli::is_dynamic_tty()) {
-    private$verbosity_level <- "minimal"
-  }
 
   super$run(func = setwd, args = list(dir = private$wd))
 
@@ -177,6 +140,11 @@ wrs_initialize <- function(
   super$run(
     func = Sys.setenv,
     args = list(WHIRL_LOG_MSG = file.path(private$wd, private$track_files_log))
+  )
+
+  saveRDS(
+    object = options(),
+    file = file.path(private$wd, "parent_options.rds")
   )
 
   environment_file <- file.path(private$wd, "_environment")
@@ -214,8 +182,7 @@ wrs_finalize <- function(self, private, super) {
 wrs_print <- function(self, private, super) {
   msg <- c(
     utils::capture.output(super$print()),
-    "Working Directory: {private$wd}",
-    "Verbose: {private$verbosity_level}"
+    "Working Directory: {private$wd}"
   )
 
   cli::cli_bullets(
@@ -228,25 +195,8 @@ wrs_print <- function(self, private, super) {
   return(invisible(self))
 }
 
-wrs_pb_update <- function(..., self, private, super) {
-  if (!is.null(private$pb)) {
-    private$pb$update(...)
-  }
-  return(invisible(self))
-}
-
-wrs_pb_done <- function(status, self, private, super) {
-  if (!is.null(private$pb)) {
-    private$pb$done(status)
-  }
-  return(invisible(self))
-}
-
 wrs_poll <- function(timeout, self, private, super) {
   status <- super$poll_process(timeout)
-  if (status == "timeout") {
-    self$pb_update()
-  }
   return(status)
 }
 
@@ -266,20 +216,43 @@ wrs_wait <- function(timeout, self, private, super) {
 wrs_check_status <- function(self, private, super) {
   status <- super$read()
   if (!is.null(status$error)) {
-    status$error |>
-      as.character() |>
-      rlang::abort()
+    cli::cli_abort(
+      c(
+        "Could not run {.file {private$current_script}}",
+        "i" = "Error from {.fn quarto::quarto_render}:",
+        strsplit(x = status$stdout, split = "\n") |>
+          unlist()
+      )
+    )
   }
   return(invisible(status))
 }
 
 wrs_log_script <- function(script, self, private, super) {
+  zephyr::msg_debug("Running script {.file {script}}")
   private$current_script <- script
 
   saveRDS(
     # Log starting time
     object = Sys.time(),
     file = file.path(private$wd, "start.rds")
+  )
+
+  saveRDS(
+    # Save packages used by the script and include the ones used to run the script
+    unique(
+      rbind(
+        renv::dependencies(
+          path = system.file("documents", "dummy.qmd", package = "whirl"),
+          quiet = TRUE
+        ),
+        renv::dependencies(
+          path = script,
+          quiet = TRUE
+        )
+      )
+    ),
+    file = file.path(private$wd, "pkgs_used.rds")
   )
 
   saveRDS(
@@ -313,16 +286,6 @@ wrs_log_script <- function(script, self, private, super) {
     )
   }
 
-  # Execute the script
-  if (private$verbosity_level != "quiet") {
-    private$pb <- pb_script$new(
-      script = private$current_script,
-      use_progress = private$verbosity_level == "verbose"
-    )
-  }
-
-  self$pb_update(status = "Running script")
-
   self$call(
     func = \(...) quarto::quarto_render(...),
     args = list(
@@ -343,7 +306,7 @@ wrs_log_script <- function(script, self, private, super) {
 }
 
 wrs_create_log <- function(self, private, super) {
-  self$pb_update(status = "Creating log")
+  zephyr::msg_debug("Creating log for {.file {private$current_script}}")
 
   if (private$track_files) {
     strace_msg <- private$wd |>
@@ -369,7 +332,6 @@ wrs_create_log <- function(self, private, super) {
       output_file = "log.html",
       execute_params = list(
         title = private$current_script,
-        approved_packages = private$approved_packages,
         with_library_paths = .libPaths(),
         track_files = private$track_files,
         tmpdir = normalizePath(private$wd)
@@ -381,16 +343,49 @@ wrs_create_log <- function(self, private, super) {
   return(invisible(self))
 }
 
-wrs_log_finish <- function(self, private, super) {
+wrs_log_finish <- function(out_dir, format, self, private, super) {
   private$result <- private$wd |>
     file.path("result.rds") |>
     readRDS()
 
-  if (!is.null(private$pb)) {
-    self$pb_done(status = private$result[["status"]][["message"]])
-  }
+  output <- wrs_create_outputs(out_dir, format, self, private, super)
 
-  return(invisible(self))
+  wrs_report_status(
+    status = output$status$message,
+    script = private$current_script,
+    logs = output$logs
+  )
+
+  invisible(output)
+}
+
+wrs_report_status <- function(status, script, logs) {
+  script_msg <- create_cli_links(
+    text = basename(script),
+    href = script
+  )
+
+  logs_msg <- create_cli_links(
+    text = get_file_ext(logs),
+    href = logs
+  )
+
+  switch(
+    EXPR = status,
+    success = zephyr::msg_success(
+      "{script_msg}: Completed succesfully. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    ),
+    warning = zephyr::msg(
+      "{script_msg}: Completed with warnings. See {cli::qty(logs_msg)}log{?s} {logs_msg}.",
+      msg_fun = cli::cli_alert_warning # Since zephyr::msg_warning only shows when verbose
+    ),
+    error = zephyr::msg_danger(
+      "{script_msg}: Completed with errors. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    ),
+    cli::cli_abort(
+      "{script}: Completed with unknown status {.emph {status}}. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    )
+  )
 }
 
 wrs_create_outputs <- function(out_dir, format, self, private, super) {