+++ NEWS.md
@@ -1,3 +1,14 @@
+# whirl 0.3.1
+
+* Improved error handling when the log cannot be created.
+* Added new option `environment_secrets` to control which secret environment variables not to include in the log.
+* Improved progress bar of `run()` to show all currently running scripts.
+* Removed `verbosity_level` argument to `run()` since it is now completely controlled by zephyr options (see `help("whirl-options")`).
+* Fixed bug where a script would not execute if using `options(warn = 2)` (#151)
+* The log now distinguishes between directly and indirectly used packages, and visually highlights their approval status if a list of approved packages is provided.
+* Added check for approved Python packages with the `approved_python_packages` option similar to for R.
+* Improved how to find used Python packages, so it now only lists the packages actually used in the script.
+
 # whirl 0.3.0
 
 * Add `write_biocompute()` to create [BioCompute Objects](https://www.biocomputeobject.org/) containing the logs in a standardized JSON format.